Rename example to CosmosDbExample
