+++ JsonApiDotNetCore.sln
@@ -44,7 +44,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MultiDbContextTests", "test
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestBuildingBlocks", "test\TestBuildingBlocks\TestBuildingBlocks.csproj", "{210FD61E-FF5D-4CEE-8E0D-C739ECCCBA21}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JsonApiDotNetCoreExample.Cosmos", "src\Examples\JsonApiDotNetCoreExample.Cosmos\JsonApiDotNetCoreExample.Cosmos.csproj", "{74819978-5C4C-460F-B5DE-31D6B6B1289C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CosmosDbExample", "src\Examples\CosmosDbExample\CosmosDbExample.csproj", "{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -212,18 +212,18 @@ Global
 		{210FD61E-FF5D-4CEE-8E0D-C739ECCCBA21}.Release|x64.Build.0 = Release|Any CPU
 		{210FD61E-FF5D-4CEE-8E0D-C739ECCCBA21}.Release|x86.ActiveCfg = Release|Any CPU
 		{210FD61E-FF5D-4CEE-8E0D-C739ECCCBA21}.Release|x86.Build.0 = Release|Any CPU
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C}.Debug|x64.Build.0 = Debug|Any CPU
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C}.Debug|x86.Build.0 = Debug|Any CPU
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C}.Release|x64.ActiveCfg = Release|Any CPU
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C}.Release|x64.Build.0 = Release|Any CPU
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C}.Release|x86.ActiveCfg = Release|Any CPU
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C}.Release|x86.Build.0 = Release|Any CPU
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}.Debug|x64.Build.0 = Debug|Any CPU
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}.Debug|x86.Build.0 = Debug|Any CPU
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}.Release|x64.ActiveCfg = Release|Any CPU
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}.Release|x64.Build.0 = Release|Any CPU
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}.Release|x86.ActiveCfg = Release|Any CPU
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -242,7 +242,7 @@ Global
 		{6CAFDDBE-00AB-4784-801B-AB419C3C3A26} = {026FBC6C-AF76-4568-9B87-EC73457899FD}
 		{EC3202C6-1D4C-4B14-A599-B9D3F27FE3BA} = {24B15015-62E5-42E1-9BA0-ECE6BE7AA15F}
 		{210FD61E-FF5D-4CEE-8E0D-C739ECCCBA21} = {24B15015-62E5-42E1-9BA0-ECE6BE7AA15F}
-		{74819978-5C4C-460F-B5DE-31D6B6B1289C} = {026FBC6C-AF76-4568-9B87-EC73457899FD}
+		{3AD5D1B4-8456-4639-98DB-F0D0EDAB2AD2} = {026FBC6C-AF76-4568-9B87-EC73457899FD}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A2421882-8F0A-4905-928F-B550B192F9A4}
