+++ src/Examples/CosmosDbExample/CosmosDbExample.csproj
