+++ src/Examples/CosmosDbExample/Definitions/TodoItemDefinition.cs
@@ -2,17 +2,17 @@ using System;
 using System.ComponentModel;
 using System.Threading;
 using System.Threading.Tasks;
+using CosmosDbExample.Models;
 using JetBrains.Annotations;
 using JsonApiDotNetCore.Configuration;
 using JsonApiDotNetCore.Middleware;
 using JsonApiDotNetCore.Queries.Expressions;
 using JsonApiDotNetCore.Resources;
-using JsonApiDotNetCoreExample.Cosmos.Models;
 using Microsoft.AspNetCore.Authentication;
 
 #pragma warning disable AV2310 // Code block should not contain inline comment
 
-namespace JsonApiDotNetCoreExample.Cosmos.Definitions
+namespace CosmosDbExample.Definitions
 {
     [UsedImplicitly(ImplicitUseKindFlags.InstantiatedNoFixedConstructorSignature)]
     public sealed class TodoItemDefinition : JsonApiResourceDefinition<TodoItem, Guid>
