+++ src/Examples/CosmosDbExample/Properties/launchSettings.json
